pull out hostdissolve functionality from discord, move a function
